[Refactor][Kernels] One CUDA-device check for every in-tree kernel
The check that a kernel's tensors are on a CUDA device existed twice: as a
helper in the pool package, which the elementwise kernels could not reach, and
as a private copy inside elementwise.py. Norm needs the same statement.

It moves to kernel_base, next to Kernel, and tolerates None so an op can pass
an absent optional input's slot straight through.
