+++ src/tileops/kernels/pool/mean_pooling.py
@@ -5,9 +5,7 @@ import tilelang
 import tilelang.language as T
 import torch
 
-from tileops.kernels.kernel_base import Kernel
-
-from .common import require_cuda
+from tileops.kernels.kernel_base import Kernel, require_cuda
 
 __all__ = ["MeanPoolingFwdKernel"]
